+++ src/lib/api/GraphQLQueries-fragments.ts
@@ -199,6 +199,8 @@ dcaCondition
     dcaVolumeMaxValue
 skipBalanceCheck
 allowRaiseToExchangeMin
+reduceToAvailableBalance
+reduceToAvailableMinSize
 baseSlOn
 closeByTimer
                     closeByTimerValue